+++ src/game/resources.rs
@@ -1,4 +1,4 @@
 use bevy::prelude::*;
 
 #[derive(Resource, Debug, Clone, Copy)]
-pub struct Lives(pub usize);
+pub struct Lives(pub u8);
